-Created seperate files for user types+changes to IUser
