+++ Admin.cs
@@ -1,0 +1,8 @@
+using FamicomBackend;
+public class Admin : IUser
+{
+	public int UserID { get; set; }
+	public string Name { get; set; }
+	public string Email { get; set; }
+
+}
