+++ Employee.cs
@@ -1,0 +1,9 @@
+using FamicomBackend;
+public class Employee : IUser
+{
+	public int UserID { get; set; }
+	public string Name { get; set; }
+	public string Email { get; set; }
+
+    public string Department { get; set; }
+}
